test(ui): consolidate Playwright suite into success + failure journeys
Finish the Stage 2 consolidation: each feature area is now one success
journey and one failure journey (~14 videos total) instead of a journey
plus scattered edge tests.

- Merge every negative/edge case per area into a single failure journey
  with a test.step() per case; each step calls mock.reset() first so
  captured requests and scenario overrides don't leak across steps.
- Fold empty states into the success journeys (chat, prompt libraries).
- Force Playwright's dev server onto webpack (--webpack). Turbopack's dev
  server corrupts the RSC client manifest when a route recompiles after a
  full navigation, breaking the second cold load of "/" (onboarding
  variants, app-shell error state). Local `npm run dev` stays on Turbopack.
- Update the README conventions + roadmap to the success/failure model.
