+++ ui/playwright/tests/prompts/prompt-libraries.spec.ts
@@ -1,26 +1,31 @@
 import { test, expect } from "../../fixtures/test";
 import { expectToast, waitForAppReady } from "../../helpers/page";
 
-// Sub-stage 2.7 — prompt libraries. /prompts redirects to ?namespace=kagent and
-// lists via GET /api/prompttemplates?namespace=<ns>. Create is a dedicated route
-// (/prompts/new) that POSTs /api/prompttemplates then redirects to the detail page.
+// Prompt libraries — one success journey + one failure journey (two videos).
+// /prompts redirects to ?namespace=kagent and lists via GET
+// /api/prompttemplates?namespace=<ns>. Create is a dedicated route (/prompts/new)
+// that POSTs /api/prompttemplates then redirects to the detail page; edit PUTs and
+// delete DELETEs it.
+//
+// Success opens on the empty namespace, then walks the library lifecycle — create,
+// view the detail page, edit fragments, delete+confirm. Failure is the single
+// name-required validation block.
 
-test.describe("prompt libraries", () => {
-  test("shows the empty state for a namespace with no libraries", async ({ page }) => {
+test("prompt library lifecycle: create, view, edit, and delete", async ({ page, mock }) => {
+  await test.step("opens on the empty state for a namespace with no libraries", async () => {
     await page.goto("/prompts");
     await waitForAppReady(page);
     await expect(page.getByRole("heading", { level: 1, name: "Prompt Libraries" })).toBeVisible();
     await expect(page.getByText("No prompt libraries in this namespace")).toBeVisible();
   });
 
-  test("creates a prompt library and POSTs the expected payload", async ({ page, mock }) => {
+  await test.step("creates a prompt library and POSTs the expected payload", async () => {
     await page.goto("/prompts/new?ns=kagent");
     await waitForAppReady(page);
 
     await page.getByLabel("Name", { exact: true }).fill("e2e-prompts");
     await page.getByLabel("Key 1").fill("safety-rules");
     await page.getByLabel("Content").fill("Always be safe.");
-
     await page.getByRole("button", { name: "Create Library" }).click();
 
     await expect(page).toHaveURL(/\/prompts\/kagent\/e2e-prompts/);
@@ -36,17 +41,14 @@ test.describe("prompt libraries", () => {
     expect(req!.body.data["safety-rules"]).toBe("Always be safe.");
   });
 
-  test("loads a library detail page", async ({ page }) => {
+  await test.step("loads the library detail page", async () => {
     await page.goto("/prompts/kagent/e2e-prompts");
     await waitForAppReady(page);
     await expect(page.getByRole("heading", { level: 1, name: "e2e-prompts" })).toBeVisible();
     await expect(page.getByRole("button", { name: "Save changes" })).toBeVisible();
   });
 
-  test("edits fragments and PUTs the update", async ({ page, mock }) => {
-    await page.goto("/prompts/kagent/e2e-prompts");
-    await waitForAppReady(page);
-
+  await test.step("edits fragments and PUTs the update", async () => {
     await page.getByLabel("Key 1").fill("safety-rules");
     await page.getByLabel("Content").fill("Always be safe.");
     await page.getByRole("button", { name: "Save changes" }).click();
@@ -60,10 +62,7 @@ test.describe("prompt libraries", () => {
     expect(req!.body.data["safety-rules"]).toBe("Always be safe.");
   });
 
-  test("deletes a library", async ({ page, mock }) => {
-    await page.goto("/prompts/kagent/e2e-prompts");
-    await waitForAppReady(page);
-
+  await test.step("deletes the library and confirms the DELETE", async () => {
     await page.getByRole("button", { name: "Delete", exact: true }).click();
     const dialog = page.getByRole("dialog");
     await expect(dialog.getByText("Delete this prompt library?")).toBeVisible();
@@ -71,18 +70,16 @@ test.describe("prompt libraries", () => {
 
     await expectToast(page, /deleted/i, { type: "success" });
     await expect(page).toHaveURL(/\/prompts\?namespace=kagent/);
-    expect(
-      await mock.lastRequest("DELETE", "/api/prompttemplates/kagent/e2e-prompts"),
-    ).not.toBeNull();
+    expect(await mock.lastRequest("DELETE", "/api/prompttemplates/kagent/e2e-prompts")).not.toBeNull();
   });
+});
 
-  test("blocks create when the name is empty", async ({ page, mock }) => {
-    await page.goto("/prompts/new?ns=kagent");
-    await waitForAppReady(page);
+test("prompt failures: blocks create when the name is empty", async ({ page, mock }) => {
+  await page.goto("/prompts/new?ns=kagent");
+  await waitForAppReady(page);
 
-    await page.getByRole("button", { name: "Create Library" }).click();
+  await page.getByRole("button", { name: "Create Library" }).click();
 
-    await expectToast(page, /Library name is required/i, { type: "error" });
-    expect((await mock.capturedRequests()).filter((r) => r.method === "POST")).toHaveLength(0);
-  });
+  await expectToast(page, /Library name is required/i, { type: "error" });
+  expect((await mock.capturedRequests()).filter((r) => r.method === "POST")).toHaveLength(0);
 });
